refactor for import availability outside of repo
