+++ chia/cmds/init_funcs.py
@@ -13,7 +13,7 @@ from chia.ssl.create_ssl import (
     generate_ca_signed_cert,
     get_chia_ca_crt_key,
     make_ca_cert,
-    write_ssl_cert_and_key,
+    write_ssl_cert_and_key, create_all_ssl,
 )
 from chia.util.bech32m import encode_puzzle_hash
 from chia.util.config import (
@@ -44,17 +44,7 @@ from chia.wallet.derive_keys import (
 )
 from chia.cmds.configure import configure
 
-_all_private_node_names: List[str] = [
-    "full_node",
-    "wallet",
-    "farmer",
-    "harvester",
-    "timelord",
-    "crawler",
-    "data_layer",
-    "daemon",
-]
-_all_public_node_names: List[str] = ["full_node", "wallet", "farmer", "introducer", "timelord", "data_layer"]
+
 
 
 def dict_add_new_default(updated: Dict, default: Dict, do_not_migrate_keys: Dict[str, Any]):
@@ -219,112 +209,6 @@ def migrate_from(
 
     return 1
 
-
-def create_all_ssl(
-    root_path: Path,
-    *,
-    private_ca_crt_and_key: Optional[Tuple[bytes, bytes]] = None,
-    node_certs_and_keys: Optional[Dict[str, Dict]] = None,
-    private_node_names: List[str] = _all_private_node_names,
-    public_node_names: List[str] = _all_public_node_names,
-    overwrite: bool = True,
-):
-    # remove old key and crt
-    config_dir = root_path / "config"
-    old_key_path = config_dir / "trusted.key"
-    old_crt_path = config_dir / "trusted.crt"
-    if old_key_path.exists():
-        print(f"Old key not needed anymore, deleting {old_key_path}")
-        os.remove(old_key_path)
-    if old_crt_path.exists():
-        print(f"Old crt not needed anymore, deleting {old_crt_path}")
-        os.remove(old_crt_path)
-
-    ssl_dir = config_dir / "ssl"
-    ca_dir = ssl_dir / "ca"
-    ensure_ssl_dirs([ssl_dir, ca_dir])
-
-    private_ca_key_path = ca_dir / "private_ca.key"
-    private_ca_crt_path = ca_dir / "private_ca.crt"
-    chia_ca_crt, chia_ca_key = get_chia_ca_crt_key()
-    chia_ca_crt_path = ca_dir / "chia_ca.crt"
-    chia_ca_key_path = ca_dir / "chia_ca.key"
-    write_ssl_cert_and_key(chia_ca_crt_path, chia_ca_crt, chia_ca_key_path, chia_ca_key, overwrite=overwrite)
-
-    # If Private CA crt/key are passed-in, write them out
-    if private_ca_crt_and_key is not None:
-        private_ca_crt, private_ca_key = private_ca_crt_and_key
-        write_ssl_cert_and_key(private_ca_crt_path, private_ca_crt, private_ca_key_path, private_ca_key)
-
-    if not private_ca_key_path.exists() or not private_ca_crt_path.exists():
-        # Create private CA
-        print(f"Can't find private CA, creating a new one in {root_path} to generate TLS certificates")
-        make_ca_cert(private_ca_crt_path, private_ca_key_path)
-        # Create private certs for each node
-        ca_key = private_ca_key_path.read_bytes()
-        ca_crt = private_ca_crt_path.read_bytes()
-        generate_ssl_for_nodes(
-            ssl_dir,
-            ca_crt,
-            ca_key,
-            prefix="private",
-            nodes=private_node_names,
-            node_certs_and_keys=node_certs_and_keys,
-            overwrite=overwrite,
-        )
-    else:
-        # This is entered when user copied over private CA
-        print(f"Found private CA in {root_path}, using it to generate TLS certificates")
-        ca_key = private_ca_key_path.read_bytes()
-        ca_crt = private_ca_crt_path.read_bytes()
-        generate_ssl_for_nodes(
-            ssl_dir,
-            ca_crt,
-            ca_key,
-            prefix="private",
-            nodes=private_node_names,
-            node_certs_and_keys=node_certs_and_keys,
-            overwrite=overwrite,
-        )
-
-    chia_ca_crt, chia_ca_key = get_chia_ca_crt_key()
-    generate_ssl_for_nodes(
-        ssl_dir,
-        chia_ca_crt,
-        chia_ca_key,
-        prefix="public",
-        nodes=public_node_names,
-        overwrite=False,
-        node_certs_and_keys=node_certs_and_keys,
-    )
-
-
-def generate_ssl_for_nodes(
-    ssl_dir: Path,
-    ca_crt: bytes,
-    ca_key: bytes,
-    *,
-    prefix: str,
-    nodes: List[str],
-    overwrite: bool = True,
-    node_certs_and_keys: Optional[Dict[str, Dict]] = None,
-):
-    for node_name in nodes:
-        node_dir = ssl_dir / node_name
-        ensure_ssl_dirs([node_dir])
-        key_path = node_dir / f"{prefix}_{node_name}.key"
-        crt_path = node_dir / f"{prefix}_{node_name}.crt"
-        if node_certs_and_keys is not None:
-            certs_and_keys = node_certs_and_keys.get(node_name, {}).get(prefix, {})
-            crt = certs_and_keys.get("crt", None)
-            key = certs_and_keys.get("key", None)
-            if crt is not None and key is not None:
-                write_ssl_cert_and_key(crt_path, crt, key_path, key)
-                continue
-
-        if key_path.exists() and crt_path.exists() and overwrite is False:
-            continue
-        generate_ca_signed_cert(ca_crt, ca_key, crt_path, key_path)
 
 
 def copy_cert_files(cert_path: Path, new_path: Path):
